# Ensure to build release binaries without dcou